fix memset overflow of wcep_pI
In the sigAllocate, the allocation size is only sizeof(picoos_int32) * PICODSP_FFTSIZE.
It will cover next cell's header to 0, which usually is int_vec29. wcep_pI is int_vec28.
And that will cause crash in picoos_deallocate, because int_vec29's header is error.
In 32 bits machine, this problem don't appear because size is covered by 0.
And in 64 bits machine, the size is covered by 0x100000000, that cause segment fault.

Signed-off-by: ganxiaolin <ganxiaolin@xiaomi.com>