Use stdin if no words given. Use stdout if no wavefile given.
Text is still not synthesized untill all text is read in,
but this is a first step.
Now pipes into pacat or aplay work (though the sapmle format
and rate need to be specified on their command line.)

For example to use pacat:
pico2wave | pacat --format=s16le --rate=8000

or to use aplay:
pico2wave | aplay -f S16_LE -r 16000

Type text to the program, then ctrl-d to signal end of file.